Added functionality to TaskList class
